pyproject.toml
[tool.poetry]
packages = [{include = "TeacherLibrary"}]

[project]
name = "TeacherLibrary"
version = "0.1.0"
description = "Mini Library Management System for English Teachers"
authors = [
    {name = "English Teachers"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.31.0",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.4",
    "openpyxl>=3.1.2",
    "requests>=2.31.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
